+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+# v2.4.0
+
+### Certificates
+
+#### Fixes
+
+* fix(certificates): Fix `inconsistent state` error on updates by including expiry and renewal params.
+* fix(certificates): Remove comma escaping logic for enrollment subject parameters.
+
+### Certificate Deployments
+
+#### Features
+* feat(deployments): Add support `overwrite` flag. *NOTE*: As of Keyfactor Command v12.0 if `overwrite=true` then the API will check if the certificate alias exists before scheduling a job.
+* feat(deployments): Add support for `redeploy` flag. Which will force a certificate to be `undeployed` and then `redeployed`.
+
+### Roles
+* chore(docs): Add deprecation notice to `keyfactor_role` resources to be replaced with `keyfactor_security_role` resource.
+* chore(docs): Add deprecation notice to `keyfactor_identity` resource to be replaced with `keyfactor_security_claim` resource.
+
 # v2.3.0
 
 ### Certificates